+++ CHANGELOG.rst
@@ -21,6 +21,8 @@ Added
 - ``raster.Image()``: expanded test_arr to test AttributeError #31
 - ``raster.Image()``: optional dimorder for arr and according test #31
 - dn2toa tests and testfiles #17
+- ``data.source()``: accept WKT string as AOI #26
+- ``data.source()``: check if an AOI string is a file or a WKT string #26
 
 Fixed
 *****